try to fix

+++ 0x15-file_io/1-create_file.c
@@ -9,7 +9,7 @@
  */
 int create_file(const char *filename, char *text_content)
 {
-	ssize_t fd, w, r;
+	int fd, w, r;
 	int i = 0;
 
 	if (!filename)
